requirements.txt
asgiref==3.5.2
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.4
django-ckeditor==6.5.1
django-cors-headers==3.13.0
django-environ==0.9.0
django-jazzmin==2.6.0
django-js-asset==2.0.0
django-modeltranslation==0.18.7
django-parler==2.3
django-rosetta==0.9.8
djangorestframework==3.14.0
idna==3.4
Pillow==9.3.0
polib==1.1.1
psycopg2-binary==2.9.5
pytz==2022.7
requests==2.28.1
sqlparse==0.4.3
typing_extensions==4.4.0
urllib3==1.26.13
drf_yasg
